feat(type): partially implement new type system
This commit adds a partial implementation of the new login system. The code for the login system is not complete and may not work correctly at this stage. Further work is needed to fully implement this

BREAKING CHANGE: This is a work in progress and should not be merged into the main branch until the it is fully implemented.